virtual_lab_eval/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.4
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
openai==2.31.0
packaging==26.0
pillow==12.2.0
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
reportlab==4.4.10
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
whitenoise==6.12.0

